Move createUser function from routes directory to controller directory
